draft-api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.136.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.13
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-classic==1.0.4
langchain-community==0.4.1
langchain-core
langchain-groq
langchain-ollama==1.1.0
langchain-openai==1.2.1
langchain-pinecone==0.2.13
langchain-protocol==0.0.12
langchain-text-splitters==1.1.2
langgraph==1.1.9
langgraph-checkpoint==4.0.2
langgraph-prebuilt==1.0.11
langgraph-sdk==0.3.13
langsmith==0.7.36
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
ollama==0.6.1
openai==2.32.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==24.2
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.4.1
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
xxhash==3.7.0
yarl==1.23.0
zstandard==0.25.0
